The proper fix for ProjectAlong for Circle
Known as giving up.
As opposed to this methods original assumptions about only having to worriy about the new value for the second radius.
Both radie can change
and the calculation for the new axies were wrong as well, (visual inspection)
And all of that sound s like a later problem as this solution works