Consistent check for linear corners
The algoritm has a special case for the points which appear on two lines due to tolerance on the corners.

The check was: If a point is within tolerance from the corner, it will be on both curve segments.

Which is false as the tolerance for the intersections act along the line, not by distance to the line.

New method calculates the distance it must be within on the current line to appear within the tolerance from the other line