Reducing risk of finding the vector getting stuck in infinite loop
THe while loop used to find the direction of the ray to be used in the IsContaining method continously mutated the checking point, leading to a relatively high risk of it getting stuck. Not mutating the point helps leviate this issue, but there is still a risk. Might require further safeguards to ensure it does not become an issue, but this initial commit should reduce the issue.