Update ProjectGraph.cs
Changeing the method from:
- DeepCloning the graph
- Looping through all entities, finding entities that are _not_ IElement0D and removing them from the dictionary and while doing that, scanning the list of relations for anyone containing the id as source or target, and if found, remove it as well.

To:
- Create a new graph, with no Entities and Relations.
- Set Entities to be all entities of type IElement0D
- Find all relations matching the GUID of one of those entities.

The above change in metodology should run drastically faster than previous interation, and generally give the expected result of a Graph with only the IElement0D types preserved, with the relations matching.

A minor change in behaviour would be for Relations that had no matching entities in the input graph. This new code will not move them on to the new graph, while the previous implementation kept them. I think this can be seen as unintentional by the previous implementation, why this refactoring should be an improvement both is speed and functionality.