Change return type of SplitAtPoints to ICurve
After discussion on PR thread with @pawelbaran agree this feels cleaner, and does not seem to have any impact on compilation of other toolkits tested for so far.

Will rely on check installer for final check of this.

Also, UTs will need to be updated again, as this means a change of the retuned items