Check SplitAtStart on main curve as well as split curve
SplitAtStart on the split segment has some tolerance issues for some cases
SplitAtStart on main curve has issues with single closed segmented curves (such as circles) hence needs to be checked for the inner split segments as well

Solution to check both. Not 100% fond of this solution, but have tried to put commetns in place as to why this is done.

Unit tests should also make sure that this is covered if further refactoring is done in the future.