Improving accuracy of SplitAtPoints for arcs
Previous implementation had issues with tolerance issues as small numeric issues built up and became larger and larger, becoming especially critical for a arc split multiple times.
New implementation remedies this issue by instead computing the angle positions for each point and then using them.
New implementation also ensures the start point of the first segment as well as the endpoint of the last segment is identical to those of the input arc.