remove unused usings to kellerman
a bit of half unrelated clean-up. Found these while searching for places where kellerman was being used, to ensure the config change was not required elsewhere, and thought I might as well remove the unused using's while I found them. Not the only file that has a reference to kellerman is ObjectDifferences.cs.