Fixes collection of dataset upgrades to match the new system
Scope to imporve this by quite a bit now that the converters live in the Verisoning_Engine. For this commit I just wanted to tweak it to work, with an as lightweight touch as possible. WIll raise an issue for this to be imporoved upon going forward, tweaking it to make use of the converters instead, now that all live in the ENgine